refactor: extract run function into testable pieces
jira: CPE-1928
