+++ src/types.ts
@@ -95,7 +95,6 @@ export interface JIRALintActionInputs {
   JIRA_TOKEN: string;
   JIRA_BASE_URL: string;
   GITHUB_TOKEN: string;
-  BRANCH_IGNORE_PATTERN: string;
   SKIP_COMMENTS: boolean;
   PR_THRESHOLD: number;
   VALIDATE_ISSUE_STATUS: boolean;
